Confidence intervals

Here we describe two methods for forming confidence intervals for empirical Bayes estimands (Empirikos.EBayesTarget).

F-Localization based intervals

Empirikos.FLocalizationInterval — Type

FLocalizationInterval(flocalization::Empirikos.FLocalization,
                      convexclass::Empirikos.ConvexPriorClass,
                      solver,
                      n_bisection = 100)

Method for computing frequentist confidence intervals for empirical Bayes estimands. Here flocalization is a Empirikos.FLocalization, convexclass is a Empirikos.ConvexPriorClass, solver is a JuMP.jl compatible solver.

n_bisection is relevant only for combinations of flocalization and convexclass for which the Charnes-Cooper transformation is not applicable. Instead, a quasi-convex optimization problem is solved by bisection and increasing n_bisection increases accuracy (at the cost of more computation).

AMARI intervals

Empirikos.AMARI — Type

AMARI(convexclass::Empirikos.ConvexPriorClass,
      flocalization::Empirikos.FLocalization,
      solver,
      plugin_G = KolmogorovSmirnovMinimumDistance(convexclass, solver))

Affine Minimax Anderson-Rubin intervals for empirical Bayes estimands. Here flocalization is a pilot Empirikos.FLocalization, convexclass is a Empirikos.ConvexPriorClass, solver is a JuMP.jl compatible solver. plugin_G is a Empirikos.EBayesMethod used as an initial estimate of the marginal distribution of the i.i.d. samples $Z$.

Interface

StatsBase.confint — Method

StatsBase.confint(method::AMARI,
                  target::Empirikos.EBayesTarget,
                  Zs;
                  α=0.05)

Form a confidence interval for the Empirikos.EBayesTargettarget with coverage 1-α based on the samples Zs using the AMARImethod.